src/Backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
bcrypt
PyJWT
